pyproject.toml
[project]
name = "openai-oauth-cli"
version = "0.1.0"
description = "CLI for OpenAI OAuth login and refresh token retrieval"
readme = "README.md"
authors = [
    { name = "Rupert-WLLP-Bai", email = "Rupert-WLLP-Bai@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "patchright>=1.52.5",
    "python-dotenv>=1.0.1",
]

[project.scripts]
openai-oauth-cli = "openai_oauth_cli.cli:main"
openai-register = "openai_register.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]

[tool.pytest.ini_options]
markers = [
    "live_e2e: runs live end-to-end flows against real external systems",
]

[tool.ruff]
target-version = "py312"

[tool.ty.environment]
python-version = "3.12"

[tool.ty.terminal]
error-on-warning = true

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
